third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/declared/get-expected.txt

This is a testharness.js-based test.
[FAIL] Getting a custom property not in the CSS rule returns undefined
  assert_equals: expected (undefined) undefined but got (object) null
[FAIL] Getting a valid property not in the CSS rule returns undefined
  assert_equals: expected (undefined) undefined but got (object) null
Harness: the test ran to completion.
